Update and rename tbq3_weekly.py to discount_code_sales.py
The input of the script is a CSV file which contains customer order information (along with the discount code used for that order if any).
The script will then evaluate the weekly total sales of all the orders that start with a specified discount i.e. "TBQ3".
The results will finally be outputted as a CSV file and saved on the specified file path.